3.1.1 PGGAN/requirements.txt
ipykernel
numpy
scipy
argparse
matplotlib
pro-gan-pth
pillow
tqdm